Clean the xml reports directory before each run:
rationale: irritatingly, JUnit5's implementation of the open xml report
standard will ONLY write to a file, not create an in-memory object. So
this is part of a hack where we pick that file out of the reports
directory and use it later on in our run. The easiest way to make sure
we're getting the latest version of the data is by wiping the previous
version.

This may bite us in the...um...KNEE...later.